requirements.txt
asgiref==3.5.2
Django==4.0.5
django-cors-headers==3.12.0
django-filter==21.1
django-rest-framework-mongoengine==3.4.1
djangorestframework==3.13.1
djongo==1.3.6
Markdown==3.3.7
mongoengine==0.24.1
pymongo==3.12.3
pytz==2022.1
sqlparse==0.2.4
